did some changes in the form

+++ frontend/src/components/ExportDetails.jsx
@@ -6,20 +6,40 @@ import { Link } from 'react-router-dom'; // Import Link from react-router-dom
 //date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
+import Swal from 'sweetalert2'; // Import SweetAlert
+
 
 const ExportDetails=({exportt})=>{
     const {dispatch}=useExportsContext()
 
-    const handleClick = async ()=>{
-        const response = await fetch('/api/export/'+exportt._id,{
-            method:'DELETE'
-        })
-        const json=await response.json()
+    const handleClick = async () => {
+        // Display confirmation dialog
+        Swal.fire({
+            title: 'Are you sure?',
+            text: 'You are about to delete this export order!',
+            icon: 'warning',
+            cancelButtonColor:'#1976D2',
+            confirmButtonColor:'#F44336',
+            showCancelButton: true,
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'No, cancel!',
+            reverseButtons: true
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                // If confirmed, proceed with deletion
+                const response = await fetch('/api/export/' + exportt._id, {
+                    method: 'DELETE'
+                });
+                const json = await response.json();
 
-        if(response.ok){
-            dispatch({type:'DELETE_EXPORT',payload:json})
-        }
-    }
+                if (response.ok) {
+                    dispatch({ type: 'DELETE_EXPORT', payload: json });
+                }
+            }
+        });
+    };
+
+    
 
     return(
         <div className="export-details">
